requirements.txt
Babel==2.7.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-crispy-forms==1.8.0
django-import-export==2.0.1
django-phonenumber-field==3.0.1
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==2.6.4
phonenumbers==8.10.22
Pillow==6.2.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.8
xlrd==1.2.0
xlwt==1.3.0
